+++ CHANGES.md
@@ -1,3 +1,4 @@
+- Cleanud up appearance menu
 - Added icon print function for php
 - Added icon shortcode