+++ CHANGES.md
@@ -1,3 +1,4 @@
+- Fixed cleanup-html-header bug
 - Fixed archive bug when retrieving body-classes
 - Cleanud up appearance menu
 - Added icon print function for php