+++ CHANGES.md
@@ -1,3 +1,5 @@
+0.1.2
+- Allow custom `content` folder (`entries` is a better name) inside `template-parts`
 - Fixed cleanup-html-header bug
 - Fixed archive bug when retrieving body-classes
 - Cleanud up appearance menu